+++ flowmachine/flowmachine/core/server/query_schemas/flowmachine_query.py
@@ -19,6 +19,8 @@ from .active_at_reference_location_counts import ActiveAtReferenceLocationCounts
 from .consecutive_trips_od_matrix import ConsecutiveTripsODMatrixSchema
 from .dummy_query import DummyQuerySchema
 from .flows import FlowsSchema
+from .outflows import OutflowsSchema
+from .inflows import InflowsSchema
 from .meaningful_locations import (
     MeaningfulLocationsAggregateSchema,
     MeaningfulLocationsBetweenLabelODMatrixSchema,
@@ -49,6 +51,8 @@ class FlowmachineQuerySchema(OneOfSchema):
     type_schemas = {
         "dummy_query": DummyQuerySchema,
         "flows": FlowsSchema,
+        "inflows": InflowsSchema,
+        "outflows": OutflowsSchema,
         "meaningful_locations_aggregate": MeaningfulLocationsAggregateSchema,
         "meaningful_locations_between_label_od_matrix": MeaningfulLocationsBetweenLabelODMatrixSchema,
         "meaningful_locations_between_dates_od_matrix": MeaningfulLocationsBetweenDatesODMatrixSchema,