+++ flowmachine/flowmachine/features/subscriber/per_contact_event_stats.py
@@ -4,10 +4,9 @@
 
 # -*- coding: utf-8 -*-
 
-from ..utilities.sets import EventsTablesUnion
 from .metaclasses import SubscriberFeature
-
-valid_stats = {"count", "sum", "avg", "max", "min", "median", "stddev", "variance"}
+from flowmachine.features.subscriber.contact_balance import ContactBalance
+from flowmachine.utils import Statistic
 
 
 class PerContactEventStats(SubscriberFeature):
@@ -22,7 +21,7 @@ class PerContactEventStats(SubscriberFeature):
     contact_balance: flowmachine.features.ContactBalance
         An instance of `ContactBalance` which lists the contacts of the
         targeted subscribers along with the number of events between them.
-    statistic : {'count', 'sum', 'avg', 'max', 'min', 'median', 'mode', 'stddev', 'variance'}, default 'avg'
+    statistic : Statistic, default Statistic.AVG
         Defaults to avg, aggregation statistic over the durations.
 
     Examples
@@ -40,16 +39,15 @@ class PerContactEventStats(SubscriberFeature):
                  ...        ...
     """
 
-    def __init__(self, contact_balance, statistic="avg"):
+    def __init__(
+        self,
+        contact_balance: ContactBalance,
+        statistic: Statistic = Statistic.AVG,
+    ):
         self.contact_balance = contact_balance
-        self.statistic = statistic
+        self.statistic = Statistic(statistic.lower())
 
-        if self.statistic not in valid_stats:
-            raise ValueError(
-                "{} is not a valid statistic. Use one of {}".format(
-                    self.statistic, valid_stats
-                )
-            )
+        super(PerContactEventStats, self).__init__()
 
     @property
     def column_names(self):
@@ -58,7 +56,7 @@ class PerContactEventStats(SubscriberFeature):
     def _make_query(self):
 
         return f"""
-        SELECT subscriber, {self.statistic}(events) AS value
+        SELECT subscriber, {self.statistic:events} AS value
         FROM ({self.contact_balance.get_query()}) C
         GROUP BY subscriber
         """