+++ secrets_quickstart/secrets-quickstart.sh
@@ -6,12 +6,18 @@
 set -eu
 set -o pipefail
 
+if [ "${BASH_VERSINFO:-0}" -le 3 ]; then
+  echo "This script requires bash version 4 or above." >/dev/stderr
+  exit 1
+fi
+
 # Ensure this docker node is in swarm mode
 docker swarm init || true
 
 
 # Remove existing stack deployment
-echo "Removing existing secrets_test_stack"
+echo "
+Removing existing secrets_test_stack"
 docker stack rm secrets_test
 
 # Wait for 'docker stack rm' to finish (see https://github.com/moby/moby/issues/30942)
@@ -37,70 +43,129 @@ if [ "$limit" -lt 0 ]; then
     exit 1
 fi
 
+# Secrets that are a random string
+rand_string_secrets=(
+  POSTGRES_PASSWORD
+  FLOWAUTH_DB_PASSWORD
+  FLOWAUTH_ADMIN_PASSWORD
+  FLOWAUTH_REDIS_PASSWORD
+  FLOWMACHINE_FLOWDB_PASSWORD
+  REDIS_PASSWORD
+  FLOWAPI_FLOWDB_PASSWORD
+  FLOWDB_POSTGRES_PASSWORD
+  FLOWETL_POSTGRES_PASSWORD
+  FLOWETL_CELERY_PASSWORD
+  FLOWETL_REDIS_PASSWORD
+  FLOWETL_WEBSERVER_PASSWORD
+  FLOWETL_SECRET_KEY
+  AIRFLOW__WEBSERVER__SECRET_KEY
+
+)
+
+#Secrets that are a random integer
+rand_int_secrets=(
+  SECRET_KEY
+)
+
+# Secrets that are hard-set here
+declare -A hard_secrets
+hard_secrets=(
+  [POSTGRES_USER]="flowdb"
+  [FLOWAUTH_ADMIN_USERNAME]="admin"
+  [FLOWMACHINE_FLOWDB_USER]="flowmachine"
+  [FLOWAPI_FLOWDB_USER]="flowapi"
+  [FLOWAPI_IDENTIFIER]="flowapi_server"
+  [FLOWETL_POSTGRES_USER]="flowetl"
+  [FLOWETL_CELERY_USER]="flowetl"
+)
+
+# Secrets that use python.cryptography to generate a Fernet key
+fernet_secrets=(
+  FLOWAUTH_FERNET_KEY
+  AIRFLOW__CORE__FERNET_KEY
+)
+
+# Secrets with specific means of generation or compounds of other secrets
+other_secrets=(
+  cert-flowkit.pem
+  key-flowkit.pem
+  PRIVATE_JWT_SIGNING_KEY
+  PUBLIC_JWT_SIGNING_KEY
+  AIRFLOW__CORE__SQL_ALCHEMY_CONN
+  AIRFLOW__CELERY__RESULT_BACKEND
+  AIRFLOW__CELERY__BROKER_URL
+  AIRFLOW_CONN_FLOWDB
+)
+
+all_secrets=(
+  "${rand_string_secrets[*]}"
+  "${rand_int_secrets[*]}"
+  "${!hard_secrets[*]}"
+  "${fernet_secrets[*]}"
+  "${other_secrets[*]}"
+)
 
 # Remove existing secrets
-echo "Removing existing secrets"
+echo "
+Removing existing secrets..."
 
-docker secret rm FLOWMACHINE_FLOWDB_PASSWORD || true
-docker secret rm FLOWMACHINE_FLOWDB_USER || true
-docker secret rm FLOWAPI_FLOWDB_PASSWORD || true
-docker secret rm FLOWAPI_FLOWDB_USER || true
-docker secret rm FLOWDB_POSTGRES_PASSWORD || true
-docker secret rm cert-flowkit.pem || true
-docker secret rm key-flowkit.pem || true
-docker secret rm REDIS_PASSWORD || true
-docker secret rm FLOWAPI_IDENTIFIER || true
-docker secret rm FLOWAUTH_ADMIN_USERNAME || true
-docker secret rm FLOWAUTH_ADMIN_PASSWORD || true
-docker secret rm FLOWAUTH_DB_PASSWORD || true
-docker secret rm PRIVATE_JWT_SIGNING_KEY || true
-docker secret rm PUBLIC_JWT_SIGNING_KEY || true
-docker secret rm FLOWAUTH_FERNET_KEY || true
-docker secret rm SECRET_KEY || true
-docker secret rm FLOWAUTH_REDIS_PASSWORD || true
-docker secret rm AIRFLOW__CORE__FERNET_KEY || true
-docker secret rm AIRFLOW__CORE__SQL_ALCHEMY_CONN || true
-docker secret rm FLOWETL_POSTGRES_PASSWORD || true
-docker secret rm AIRFLOW_CONN_FLOWDB || true
+for secret in ${all_secrets[*]} ; do
+    docker secret rm $secret || true
+done
 
-# Add new secrets
+# Each of these loops:
+# -Generates the relevent type of secret
+# -Exports the name and value of the secret to the local environment
+# By the end, each secret should be in docker and the local env
 
-echo "Adding secrets"
-# Flowauth
-openssl genrsa -out tokens-private-key.key 4096
-FLOWAUTH_DB_PASSWORD=$(openssl rand -base64 16 | tr -cd '0-9-a-z-A-Z')
-echo "$FLOWAUTH_DB_PASSWORD" | docker secret create FLOWAUTH_DB_PASSWORD -
-FLOWAUTH_ADMIN_PASSWORD=$(openssl rand -base64 16 | tr -cd '0-9-a-z-A-Z')
-echo "$FLOWAUTH_ADMIN_PASSWORD"| docker secret create FLOWAUTH_ADMIN_PASSWORD -
-FLOWAUTH_REDIS_PASSWORD=$(openssl rand -base64 16 | tr -cd '0-9-a-z-A-Z')
-echo "$FLOWAUTH_REDIS_PASSWORD"| docker secret create FLOWAUTH_REDIS_PASSWORD -
-SECRET_KEY=$(openssl rand -base64 64)
-echo "$SECRET_KEY"| docker secret create SECRET_KEY -
-echo "admin" | docker secret create FLOWAUTH_ADMIN_USERNAME -
+echo "
+Generating random string secrets..."
+for secret_name in ${rand_string_secrets[*]} ; do
+  echo "Generating $secret_name"
+  declare "${secret_name}"="$(openssl rand -base64 16 | tr -cd '0-9-a-z-A-Z')"
+  echo "${!secret_name}" | docker secret create "${secret_name}" -
+done
+
+echo "
+Generating random int secrets..."
+for secret_name in ${rand_int_secrets[*]} ; do
+  echo "Generating $secret_name"
+  declare "${secret_name}"="$(openssl rand -base64 64)"
+  echo "${!secret_name}" | docker secret create "${secret_name}" -
+done
+
+echo "
+Setting hard-coded secrets..."
+for secret_name in ${!hard_secrets[*]} ; do
+  declare "${secret_name}"="${hard_secrets[${secret_name}]}"
+  echo "Setting ${secret_name} to ${!secret_name}"
+  echo "${!secret_name}" | docker secret create "${secret_name}" -
+done
+
+echo "
+Setting up Fernet key generation..."
 pip install cryptography
-FLOWAUTH_FERNET_KEY=$(python -c "from cryptography.fernet import Fernet; print(Fernet.generate_key().decode())")
-echo "$FLOWAUTH_FERNET_KEY"| docker secret create FLOWAUTH_FERNET_KEY -
 
+echo "
+Generating Fernet keys..."
+for secret_name in ${fernet_secrets[*]} ; do
+  echo "Generating $secret_name"
+  declare "${secret_name}"="$(python3 -c "from cryptography.fernet import Fernet; print(Fernet.generate_key().decode())")"
+  echo "${!secret_name}" | docker secret create "${secret_name}" -
+done
 
-# Flowmachine
-FLOWMACHINE_FLOWDB_PASSWORD=$(openssl rand -base64 16 | tr -cd '0-9-a-z-A-Z')
-echo "$FLOWMACHINE_FLOWDB_PASSWORD" | docker secret create FLOWMACHINE_FLOWDB_PASSWORD -
-echo "flowmachine" | docker secret create FLOWMACHINE_FLOWDB_USER -
-echo "flowapi" | docker secret create FLOWAPI_FLOWDB_USER -
-REDIS_PASSWORD=$(openssl rand -base64 16 | tr -cd '0-9-a-z-A-Z')
-echo "$REDIS_PASSWORD"| docker secret create REDIS_PASSWORD -
+echo "
+Generating Flowauth RSA key"
+openssl genrsa -out tokens-private-key.key 4096
 
-# FlowDB
-FLOWAPI_FLOWDB_PASSWORD=$(openssl rand -base64 16 | tr -cd '0-9-a-z-A-Z')
-echo "$FLOWAPI_FLOWDB_PASSWORD" | docker secret create FLOWAPI_FLOWDB_PASSWORD -
-FLOWDB_POSTGRES_PASSWORD=$(openssl rand -base64 16 | tr -cd '0-9-a-z-A-Z')
-echo "$FLOWDB_POSTGRES_PASSWORD" | docker secret create FLOWDB_POSTGRES_PASSWORD -
-
-# FlowAPI
+echo "
+Generating FlowAPI public-private key pair"
 openssl rsa -pubout -in  tokens-private-key.key -out tokens-public-key.pub
 docker secret create PRIVATE_JWT_SIGNING_KEY tokens-private-key.key
 docker secret create PUBLIC_JWT_SIGNING_KEY tokens-public-key.pub
-echo "flowapi_server" | docker secret create FLOWAPI_IDENTIFIER -
+
+echo "
+Generating FlowAPI SSL cert"
 openssl req -newkey rsa:4096 -days 3650 -nodes -x509 -subj "/CN=flow.api" \
     -extensions SAN \
     -config <( \
@@ -118,15 +183,23 @@ fi
 docker secret create cert-flowkit.pem cert-flowkit.pem
 docker secret create key-flowkit.pem key-flowkit.pem
 
-# FlowETL
-AIRFLOW__CORE__FERNET_KEY=$(python -c "from cryptography.fernet import Fernet; print(Fernet.generate_key().decode())")
-echo "AIRFLOW__CORE__FERNET_KEY"| docker secret create AIRFLOW__CORE__FERNET_KEY -
+echo "
+Generating compound secrets..."
+echo "Setting AIRFLOW__CORE__SQL_ALCHEMY_CONN"
+AIRFLOW__CORE__SQL_ALCHEMY_CONN="postgresql://$FLOWETL_POSTGRES_USER:$FLOWETL_POSTGRES_PASSWORD@flowetl_db:5432/flowetl"
+echo "${AIRFLOW__CORE__SQL_ALCHEMY_CONN}" | docker secret create AIRFLOW__CORE__SQL_ALCHEMY_CONN -
+echo "Setting AIRFLOW__CELERY__RESULT_BACKEND"
+AIRFLOW__CELERY__RESULT_BACKEND="db+${AIRFLOW__CORE__SQL_ALCHEMY_CONN:?}"
+echo "$AIRFLOW__CELERY__RESULT_BACKEND" | docker secret create AIRFLOW__CELERY__RESULT_BACKEND -
+echo "Setting AIRFLOW__CELERY__BROKER_URL"
+AIRFLOW__CELERY__BROKER_URL="redis://:${FLOWETL_REDIS_PASSWORD:?}@flowetl_redis:6379/0"
+echo "$AIRFLOW__CELERY__BROKER_URL" | docker secret create AIRFLOW__CELERY__BROKER_URL -
+echo "AIRFLOW_CONN_FLOWDB"
+AIRFLOW_CONN_FLOWDB="postgresql://$POSTGRES_USER:$POSTGRES_PASSWORD@flowdb:5432/flowdb"
+echo "$AIRFLOW_CONN_FLOWDB" | docker secret create AIRFLOW_CONN_FLOWDB -
 
-FLOWETL_POSTGRES_PASSWORD=$(openssl rand -base64 16 | tr -cd '0-9-a-z-A-Z')
-echo "$FLOWETL_POSTGRES_PASSWORD" | docker secret create FLOWETL_POSTGRES_PASSWORD -
-AIRFLOW__CORE__SQL_ALCHEMY_CONN="postgres://flowetl:$FLOWETL_POSTGRES_PASSWORD@flowetl_db:5432/flowetl"
-echo "AIRFLOW__CORE__SQL_ALCHEMY_CONN" | docker secret create AIRFLOW__CORE__SQL_ALCHEMY_CONN -
-echo "postgres://flowdb:$FLOWDB_POSTGRES_PASSWORD@flowdb:5432/flowdb" | docker secret create AIRFLOW_CONN_FLOWDB -
+echo "Secret gen complete"
+
 export FLOWETL_HOST_GROUP_ID=$(id -g)
 export FLOWETL_HOST_USER_ID=$(id -u)
 
@@ -138,7 +211,6 @@ export FLOWETL_HOST_PORT=8080
 export REDIS_HOST_PORT=6379
 
 # Bind mounts
-
 export FLOWDB_HOST_GROUP_ID=$(id -g)
 export FLOWDB_HOST_USER_ID=$(id -u)
 export FLOWDB_DATA_DIR=./flowdb_pgdata
@@ -186,6 +258,9 @@ $FLOWAUTH_REDIS_PASSWORD
 FlowMachine redis password
 $REDIS_PASSWORD
 
+FlowETL redis password
+$FLOWETL_REDIS_PASSWORD
+
 FlowETL Fernet Key
 $AIRFLOW__CORE__FERNET_KEY
 
@@ -195,6 +270,13 @@ $AIRFLOW__CORE__SQL_ALCHEMY_CONN
 FlowDB data dir
 $FLOWDB_DATA_DIR
 
+Airflow flowdb connection
+$AIRFLOW_CONN_FLOWDB
+
+Airflow webserver user
+admin
+
+Airflow webserver password
+$FLOWETL_WEBSERVER_PASSWORD
 ===============
 "
-